filtered negative samples to exclude day 1 and dialysis patients

+++ sql_queries/make_dataset/create_load_select_tables.sql
@@ -94,7 +94,7 @@ CREATE TABLE intakeoutput
 (
 	intakeoutputid INT NOT NULL,
     patientunitstayid INT NOT NULL,
-	intakeoutputoffset INT NOT NULL,
+	intakeoutputoffset FLOAT NOT NULL,
 	intaketotal NUMERIC(12,4),
 	outputtotal NUMERIC(12,4),
 	dialysistotal NUMERIC(12,4),
